refactor: just moved all the components representing a page in a "page" folder
