+++ client/src/app/app.module.ts
@@ -9,15 +9,15 @@ import { itemsReducer } from './reducers/items.reducer';
 import { AppComponent } from './app.component';
 import { NavbarComponent } from './components/navbar/navbar.component';
 import { AppRoutingModule } from './app-routing.module';
-import { LoginComponent } from './components/login/login.component';
-import { LoadingComponent } from './components/loading/loading.component';
+import { LoginComponent } from './components/pages/login/login.component';
+import { LoadingComponent } from './components/pages/loading/loading.component';
 import { ItemCardSearchComponent } from './components/item-card-search/item-card-search.component';
-import { SearchPageComponent } from './components/search-page/search-page.component';
-import { BorrowPageComponent } from './components/borrow-page/borrow-page.component';
+import { SearchPageComponent } from './components/pages/search-page/search-page.component';
+import { BorrowPageComponent } from './components/pages/borrow-page/borrow-page.component';
 import { ItemCardBorrowComponent } from './components/item-card-borrow/item-card-borrow.component';
-import { BorrowPageUsersComponent } from './components/borrow-page-users/borrow-page-users.component';
+import { BorrowPageUsersComponent } from './components/pages/borrow-page-users/borrow-page-users.component';
 import { UserCardComponent } from './components/user-card/user-card.component';
-import { BorrowPageLendComponent } from './components/borrow-page-lend/borrow-page-lend.component';
+import { BorrowPageLendComponent } from './components/pages/borrow-page-lend/borrow-page-lend.component';
 
 @NgModule({
   declarations: [
