+++ client/src/app/components/pages/borrow-page-lend/borrow-page-lend.component.html
