+++ client/src/app/components/pages/loading/loading.component.scss
