+++ client/src/app/components/pages/login/login.component.scss
@@ -1,4 +1,4 @@
-@import "../../../variables.scss";
+@import "../../../../variables.scss";
 
 .form {
   padding: 20rem;
