+++ client/src/app/components/pages/search-page/search-page.component.ts
